requirements.txt
pandas
openpyxl
Pillow
opencv-python
ffpyplayer
requests
